feat: add SDID (Synthetic Difference-in-Differences) econometric method
基于 Arkhangelsky et al. (2021)，实现合成双重差分计量模型：

## 算法实现
- 单元权重 omega：cvxpy 凸优化求解，含截距 r-hat 以匹配水平差异
- 时间权重 lambda：cvxpy 回归求解，平衡预处理与后处理期信息
- 加权 DID：逐期效应 tau_t = (处理组 - 合成控制) - lambda 加权基线
- Placebo 方差推断：对每个控制单元运行 SDID 以估计标准误
- 控制变量增强：标准化协变量均值作为优化增强行
- 正则化参数 zeta 自动估计与缓存复用

## 关键设计
- 支持多处理单元（取均值）
- zeta/zeta_time 缓存确保 Conformal 重拟合中模型一致性
- 缓存守卫防止截断数据污染完整数据缓存
- 求解器失败/权重退化时回退等权重并警告
- 预处理拟合 RMSE 诊断
- Placebo 计算含 tqdm 进度条

## 调用方修改
- base.py: _econ_fit() 透传 **kwargs 至 fit_econmodel()
- base.py: conformal_inference() 传递 **kwargs 至 _econ_fit()

## 测试
- 11 项集成测试全部通过
- 原假设精确偏移验证（误差 < 1e-10）
- zeta 一致性验证（基准 vs Conformal 重拟合）
- 边缘情况覆盖（单处理组、全处理、过少预处理期）

+++ src/econformal/conformal_methods/full.py
@@ -109,7 +109,7 @@ class Conformal(ConformalBase):
     def predict(self):
         """为每个新样本计算预测区间"""
         # 生成nulls矩阵,（行为p_value_matrix行数，列为nulls长度，每行都是nulls，方便找区间）
-        nulls_matrix = np.tile(self.nulls, (self.p_value_matrix.shape[0], 1))
+        nulls_matrix = np.tile(self.nulls, (self.p_value_matrix.shape[0], 1)).astype(float)
 
         # 找到self.p_value_matrix中p值小于1-coverage的位置，并将对应nulls设为NaN
         mask = self.p_value_matrix < (1-self.coverage)
